web-server/requirements.txt
asgiref==3.8.1
Django==5.1.1
djangorestframework==3.15.2
grpcio==1.66.1
grpcio-tools==1.66.1
pika==1.3.2
protobuf==5.28.2
psycopg2-binary==2.9.9
setuptools==75.1.0
sqlparse==0.5.1
